Finished all product test

+++ src/controllers/product.js
@@ -14,11 +14,13 @@ async function getAllProducts(req, res) {
 
     const dataCategory = await Category.findById(category);
 
+    if (dataCategory == undefined) return res.status(500).json({ error: 'Category not founded' });
+
     if (category != undefined && name == undefined) products = await Promise.all(dataCategory.products.map(async product => await Product.findById(product)));
 
     if (name != undefined && category == undefined) products = await Promise.all(await Product.find({ "name": { $regex: name, $options: 'i' } }));
 
-    if (category && name) return res.status(400).json({ error: 'only supports one query at a time' });
+    if (category && name) return res.status(500).json({ error: 'only supports one query at a time' });
 
     res.json({ products });
   } catch (error) {
@@ -33,7 +35,7 @@ async function getProduct(req, res) {
 
   try {
     const product = await Product.findById(product_id);
-    if (product == undefined) return res.status(400).json({ error: 'Product not founded' });
+    if (product == undefined) return res.status(500).json({ error: 'Product not founded' });
 
     res.json({ product });
   } catch (error) {
@@ -49,12 +51,13 @@ async function createProduct(req, res) {
   if (!name || !price || !seller || !status || !category) return res.status(400).json({ error: 'Missing product data' });
 
   try {
-    const product = await Product.create({ name, price, seller, status, category });
 
     const dataUser = await User.findById(seller);
     const dataCategory = await Category.findById(category);
 
     if (dataUser == undefined || dataCategory == undefined) return res.status(500).json({ error: "User or category not founded" })
+
+    const product = await Product.create({ name, price, seller, status, category });
 
     await Category.findByIdAndUpdate(category, { $push: { 'products': product } });
     await User.findByIdAndUpdate(seller, { $push: { 'productsOnSell': product } });
@@ -67,9 +70,9 @@ async function createProduct(req, res) {
 
 async function updateProduct(req, res) {
   const { product_id } = req.params;
-  const { name, price, rating, seller, category } = req.body;
+  const { name, price, rating, seller, category, status } = req.body;
   if (!product_id) return res.status(400).json({ error: 'Missing product_id' });
-  if (!name && !price && !rating && !seller && !category) return res.status(400).json({ error: 'Missing product data' });
+  if (!name && !price && !rating && !seller && !category && !status) return res.status(400).json({ error: 'Missing product data' });
 
   try {
     const product = await Product.findById(product_id);
@@ -81,7 +84,8 @@ async function updateProduct(req, res) {
       price: price || product.price,
       rating: rating || product.rating,
       seller: seller || product.seller,
-      category: category || product.category
+      category: category || product.category,
+      status: status || product.status
     }
 
     const dataUser = await User.findById(data.seller);
@@ -90,7 +94,7 @@ async function updateProduct(req, res) {
     if (dataUser == undefined || dataCategory == undefined) return res.status(500).json({ error: "User or category not founded" })
 
     const newProduct = await Product.findByIdAndUpdate(product_id, data, { new: true });
-    res.json({ newProduct });
+    res.json({ product: newProduct });
   } catch (error) {
     res.status(400).json({ error: 'Invalid product_id or data' });
   }
@@ -102,24 +106,24 @@ async function deleteProduct(req, res) {
 
   try {
     const product = await Product.findById(product_id);
+
+    if (product == undefined) return res.status(500).json({ error: 'Product not founded' });
+
     const user = await User.findById(product.seller);
     user.productsOnSell = user.productsOnSell.filter(product => product != product_id);
-
     await User.findByIdAndUpdate(product.seller, user, { new: true })
 
     const category = await Category.findById(product.category);
     category.products = category.products.filter(product => product != product_id);
-
     await Category.findByIdAndUpdate(product.category, category, { new: true })
 
     product.reviews.map(async review => await Review.findByIdAndDelete(review));
-
 
     await Product.findByIdAndDelete(product_id);
 
     res.json({ product });
   } catch (error) {
-    res.status(400).json({ error: 'Invalid product_id' });
+    res.status(500).json({ error: 'Invalid product_id' });
   }
 }
 
@@ -129,7 +133,7 @@ async function getUserProducts(req, res) {
 
   try {
     const user = await User.findById(user_id);
-    if (user == undefined) return res.status(400).json({ error: 'User not founded' });
+    if (user == undefined) return res.status(500).json({ error: 'User not founded' });
 
     res.json({ products: user.productsOnSell });
   } catch (error) {
